chore: address #279 review — move Capability accessor traits into preludes
Field accessors on Capability<Effect> were defined alongside their
effect type with the *Capability suffix:

- ResolveCapability, PublishCapability, RetractCapability in memory.rs
- GetCapability, PutCapability in archive.rs

Moves them into the corresponding prelude module and renames to *Ext
to match the existing naming convention there (MemorySubjectExt,
MemoryExt, SpaceExt, CellExt, ArchiveSubjectExt, ArchiveExt, CatalogExt):

- ResolveExt, PublishExt, RetractExt in memory/prelude.rs
- GetExt, PutExt in archive/prelude.rs

Consumers in dialog-storage providers updated to import the new traits
from the prelude.
