+++ rust/dialog-storage/src/storage/provider/indexeddb/memory.rs
@@ -7,10 +7,8 @@ use super::{IndexedDb, to_uint8array};
 use async_trait::async_trait;
 use dialog_capability::{Capability, Provider};
 use dialog_common::Blake3Hash;
-use dialog_effects::memory::{
-    Edition, MemoryError, Publish, PublishCapability, Resolve, ResolveCapability, Retract,
-    RetractCapability, Version,
-};
+use dialog_effects::memory::prelude::{PublishExt, ResolveExt, RetractExt};
+use dialog_effects::memory::{Edition, MemoryError, Publish, Resolve, Retract, Version};
 use js_sys::Uint8Array;
 use wasm_bindgen::{JsCast, JsValue};
